# openapi/control-tower/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Genesis Control Tower API
  version: 1.0.0
  description: Governed fleet operations, command authorization, incident management and evidence publication.
security:
  - bearerAuth: []
    mutualTLS: []
paths:
  /v1/control-tower/fleet/refresh:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FleetRefreshRequest'}}}}
      responses: {'200': {description: Fleet refreshed, content: {application/json: {schema: {$ref: '#/components/schemas/FleetRefreshReceipt'}}}}}
  /v1/control-tower/fleet:
    get:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: status, in: query, schema: {$ref: '#/components/schemas/FleetStatus'}}]
      responses: {'200': {description: Fleet snapshots, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/FleetNodeSnapshot'}}}}}}
  /v1/control-tower/nodes/{node_id}:
    get:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: node_id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Node snapshot, content: {application/json: {schema: {$ref: '#/components/schemas/FleetNodeSnapshot'}}}}}
  /v1/control-tower/command-requests:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateCommandRequest'}}}}
      responses: {'201': {description: Authorization request created, content: {application/json: {schema: {$ref: '#/components/schemas/ControlCommandRequest'}}}}}
  /v1/control-tower/command-requests/{request_id}/authorize:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: request_id, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AuthorizeCommandRequest'}}}}
      responses: {'200': {description: Authorization recorded, content: {application/json: {schema: {$ref: '#/components/schemas/ControlCommandRequest'}}}}}
  /v1/control-tower/command-requests/{request_id}/dispatch:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: request_id, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DispatchCommandRequest'}}}}
      responses: {'200': {description: Command dispatched, content: {application/json: {schema: {$ref: '#/components/schemas/ControlCommandRequest'}}}}}
  /v1/control-tower/incidents:
    get:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: status, in: query, schema: {type: string}}]
      responses: {'200': {description: Incidents, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ControlIncident'}}}}}}
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateIncidentRequest'}}}}
      responses: {'201': {description: Incident created, content: {application/json: {schema: {$ref: '#/components/schemas/ControlIncident'}}}}}
  /v1/control-tower/incidents/{incident_id}/transition:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: incident_id, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IncidentTransitionRequest'}}}}
      responses: {'200': {description: Incident transitioned, content: {application/json: {schema: {$ref: '#/components/schemas/ControlIncident'}}}}}
  /v1/control-tower/publications/lease:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PublicationLeaseRequest'}}}}
      responses: {'200': {description: Publication lease, content: {application/json: {schema: {$ref: '#/components/schemas/PublicationLease'}}}}}
  /v1/control-tower/publications/{lease_id}/ack:
    post:
      security: [{bearerAuth: [], mutualTLS: []}]
      parameters: [{name: lease_id, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PublicationAckRequest'}}}}
      responses: {'200': {description: Publication acknowledged, content: {application/json: {schema: {$ref: '#/components/schemas/PublicationReceipt'}}}}}
  /v1/control-tower/dashboard:
    get:
      security: [{bearerAuth: [], mutualTLS: []}]
      responses: {'200': {description: Operational dashboard, content: {application/json: {schema: {$ref: '#/components/schemas/DashboardSummary'}}}}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
    mutualTLS: {type: mutualTLS}
  schemas:
    FleetStatus: {type: string, enum: [ONLINE, DEGRADED, OFFLINE, QUARANTINED, DRAINING, RETIRED, UNKNOWN]}
    FleetRefreshRequest: {type: object, additionalProperties: false, required: [actor_id], properties: {actor_id: {type: string}}}
    FleetRefreshReceipt: {type: object, required: [observed_at, refreshed_node_ids, opened_incident_ids, resolved_incident_ids, evidence_event_id], properties: {observed_at: {type: string, format: date-time}, refreshed_node_ids: {type: array, items: {type: string}}, opened_incident_ids: {type: array, items: {type: string}}, resolved_incident_ids: {type: array, items: {type: string}}, evidence_event_id: {type: string}}}
    FleetNodeSnapshot: {type: object, required: [node_id, effective_status, observed_at], properties: {node_id: {type: string}, effective_status: {$ref: '#/components/schemas/FleetStatus'}, observed_at: {type: string, format: date-time}}}
    CreateCommandRequest: {type: object, required: [request_id, node_id, command_type, requested_by, target_reference, purpose, expires_at], properties: {request_id: {type: string}, node_id: {type: string}, command_type: {type: string}, requested_by: {type: string}, target_reference: {type: string}, payload: {type: object}, purpose: {type: string}, expires_at: {type: string, format: date-time}}}
    AuthorizeCommandRequest: {type: object, required: [decision, actor_id, reason], properties: {decision: {type: string, enum: [APPROVE, REJECT]}, actor_id: {type: string}, reason: {type: string}, capability_id: {type: [string, 'null']}}}
    DispatchCommandRequest: {type: object, required: [actor_id], properties: {actor_id: {type: string}}}
    ControlCommandRequest: {type: object, required: [request_id, node_id, status, expires_at], properties: {request_id: {type: string}, node_id: {type: string}, status: {type: string}, expires_at: {type: string, format: date-time}, dispatched_command_id: {type: [string, 'null']}}}
    CreateIncidentRequest: {type: object, required: [incident_id, node_id, severity, incident_type, summary, opened_by], properties: {incident_id: {type: string}, node_id: {type: string}, severity: {type: string}, incident_type: {type: string}, summary: {type: string}, opened_by: {type: string}}}
    IncidentTransitionRequest: {type: object, required: [status, actor_id, note], properties: {status: {type: string}, actor_id: {type: string}, note: {type: string}}}
    ControlIncident: {type: object, required: [incident_id, node_id, severity, status], properties: {incident_id: {type: string}, node_id: {type: string}, severity: {type: string}, status: {type: string}}}
    PublicationLeaseRequest: {type: object, required: [publisher_id], properties: {publisher_id: {type: string}, max_events: {type: integer, minimum: 1, maximum: 100}, lease_seconds: {type: [integer, 'null']}}}
    PublicationLease: {type: object, required: [lease_id, publisher_id, lease_token, events, expires_at], properties: {lease_id: {type: string}, publisher_id: {type: string}, lease_token: {type: string}, events: {type: array, items: {type: object}}, expires_at: {type: string, format: date-time}}}
    PublicationAckRequest: {type: object, required: [publisher_id, lease_token, items], properties: {publisher_id: {type: string}, lease_token: {type: string}, items: {type: array, minItems: 1, items: {type: object}}}}
    PublicationReceipt: {type: object, required: [lease_id, acknowledged_event_ids, acknowledged_at], properties: {lease_id: {type: string}, acknowledged_event_ids: {type: array, items: {type: string}}, acknowledged_at: {type: string, format: date-time}}}
    DashboardSummary: {type: object, required: [generated_at, total_nodes, nodes_by_status, open_incidents, critical_incidents, authorization_queue, authorized_commands, unpublished_evidence], properties: {generated_at: {type: string, format: date-time}, total_nodes: {type: integer}, nodes_by_status: {type: object}, open_incidents: {type: integer}, critical_incidents: {type: integer}, authorization_queue: {type: integer}, authorized_commands: {type: integer}, unpublished_evidence: {type: integer}}}